Test algorithm on 3x3 game for simplicity
The rewards have been adjusted to be 1 (as final reward) in case of victory, 0 in case of loss. Negative rewards are causing overfitting. As a result, the algorithm only learns how to play player 1. Moreover half of the samples are useless as the player who loses has a loss of 0.

Still, performance is decent (>90% against a random opponent).

Same approach on 2x2 converges to the optimal strategy after a couple of episodes. This was expected as the optimal policy for player 1 in the 2x2 case is constant.