Extract policy update function as parameter
The former reinforce(...) function is now called policy_gradient(...). reinforce(...) is now a smaller function only responsible of computing the loss and updating the policy network. This is provided to the policy_gradient(...) function via a higher order parameter, making the approach more expandible and general. This enables implementation of the PPO algorithm via potentially little effort.