Extract advantage function computation
Still not added as parameter, but ready to do so. Some shenanigans are necessary in order to keep an episode wise subdivision of the states/rewards during multiple episodes. This is necessary as the advantage functions must be computed at episode level, often times having direct access to individual rewards (e.g. GAE).