Add reproducibility mechanism everywhere
Full reproducibility is now obtainable. This is achieved via seed/rng parameters wherever necessary (in sampling functions, players, training procedures, ...). However, note that: NN initialization (i.e. construction of a PolicyNetwork or ValueFunctionNetwork) will by default lead to random initialization. To ensure reproducibility of this step, it will be necessary to set torch's global seed (discouraged) or initialize the models within a torch.fork_rng() context, with fixed seed (encouraged).

Moreover, the random sequence used for mid-train evaluation is decoupled from the training process, ensuring reproducible results of the train process independently of how and if they are tested. A single run (train) on the 3x3 grid will take ~10min (11th Gen Intel(R) Core(TM) i7-1195G7 @ 2.90GHz), but roughly 10-20% of the total time is spent evaluating with my setup (hundreds of evaluation games on minmax with different depths and random player). This time is even higher for larger spaces with a higher branching factor (5x5 grid and more), so skipping evaluation can be a viable option to reduce train time when recomputing a result using a known seed.