Add support for device selection to AZ
Still, the single naive single threaded nature of the implementation make this potentially inefficient. A lot of transfers between the devices may happen e.g. between cpu and gpu when using a cuda device. A viable fix would be proper multithreading implementation of episode generation, accounting for batched computation.

Still, the current approach is valid in a cuda scenario and will get decent results for the 5x5 game withing a few days (on a colab-like environment).