Fix temperature ignored by AZPlayer
Would always default to 1. This makes previous evaluations with zero temperature unfair towards the az agent, since with a temperature of 1 there is still a decent probability of choosing a randomly losing action. New evaluations with a temperature of 0 show 100% winning scenarios against all MM agents up to depth 3 as both player1 and player2, with any temperature.